1.5.19 * Android: rework full-screen layout to account for Android 4.0 behavior * iOS: add UITextField to main view, not window, so it rotates properly when using GLKView * Move some hq2x calculations to vertex shader * Support key repeats on iCade * Disable key repeats in EmuFramework when assigning keys or detecting input device * Various cleanup & refactoring * Build tested on Android, desktop Linux, & iOS

+++ imagine/src/base/android/androidBase.cc
@@ -375,7 +375,6 @@ static bool setOrientationOS(int o)
 		{
 			auto rotAngle = orientationDiffTable[osOrientation][o];
 			logMsg("animating from %d degrees", (int)angleToDegree(rotAngle));
-			//projAngleM.set(IG::toRadians(rotAngle), 0., INTERPOLATOR_TYPE_EASEOUTQUAD, 10);
 			Gfx::animateProjectionMatrixRotation(rotAngle, 0.);
 		}
 		//logMsg("new value %d", o);
@@ -1211,7 +1210,7 @@ static void onInputQueueCreated(ANativeActivity* activity, AInputQueue* queue)
 		[](int, int, void* data)
 		{
 			processInput((AInputQueue*)data);
-			return (int)1;
+			return 1;
 		}, queue);
 }
